+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
 ## Pre release
+
 ### Enhancements
+- GLS-458 - Return dataset source data as metadata in the response
+
 ### Bugs fixed
 - GLS-428 - Handle missing markets from the IMF dataset
 - GLS-425 - Fix trade in services queryset
+
 ## [2.5.0](https://github.com/uktrade/directory-api/releases/tag/2.5.0)
 [Full Changelog](https://github.com/uktrade/directory-api/compare/2.4.2...2.5.0)
 ### Enhancements